requirements.txt
requests==2.23.0
python-dotenv==0.12.0requests==2.23.0
requests-unixsocket==0.1.5
